+++ ski_jumping.dto/src/main/java/com.pl.skijumping.dto/DataRaceDTO.java
@@ -16,5 +16,5 @@ public class DataRaceDTO {
     private String shortCountryName;
     private String competitionName;
     private String competitionType;
-    private Integer raceId;
+    private Long raceId;
 }
